requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
